+++ pcs/cli/routing/prop.py
@@ -1,8 +1,5 @@
-from pcs import (
-    prop,
-    usage,
-)
-from pcs.cli import cluster_property
+from pcs import usage
+from pcs.cli.cluster_property import command as cluster_property
 from pcs.cli.common.routing import create_router
 
 property_cmd = create_router(
@@ -12,13 +9,15 @@ property_cmd = create_router(
         "unset": cluster_property.unset_property,
         # TODO remove, deprecated command
         # replaced with 'config'
-        "list": prop.list_property_deprecated,
+        "list": cluster_property.list_property_deprecated,
         # TODO remove, deprecated command
         # replaced with 'config'
-        "show": prop.list_property_deprecated,
-        "config": prop.list_property,
+        "show": cluster_property.list_property_deprecated,
+        "config": cluster_property.config,
+        "defaults": cluster_property.defaults,
+        "describe": cluster_property.describe,
         "get_cluster_properties_definition": (
-            prop.print_cluster_properties_definition
+            cluster_property.print_cluster_properties_definition_legacy
         ),
     },
     ["property"],