+++ pcs/common/pacemaker/cluster_property.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import Sequence
+
+from pcs.common.interface.dto import DataTransferObject
+from pcs.common.resource_agent.dto import ResourceAgentParameterDto
+from pcs.common.types import StringCollection
+
+
+@dataclass(frozen=True)
+class ClusterPropertyMetadataDto(DataTransferObject):
+    properties_metadata: Sequence[ResourceAgentParameterDto]
+    readonly_properties: StringCollection